refactor(compile): extract module loader
